backend/requirements.txt
asgiref==3.9.2
Django==5.2.7
sqlparse==0.5.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
channels==4.1.0
psycopg2-binary==2.9.11
supabase==2.23.0
requests ==2.31.0
